refactor: change folder structures
