+++ package.json
@@ -62,6 +62,6 @@
     }
   },
   "dependencies": {
-    "@yuno-payments/sdk-web": "^2.0.0"
+    "@yuno-payments/sdk-web": "2.0.0"
   }
 }
