+++ src/domain/payment/index.ts
@@ -1,2 +1,2 @@
-export * from "./checkout";
+export * from "./checkout/wallet/index";
 export * from "./token";
